# shiftory-server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shiftory API
  version: 1.0.0
  description: 排班协同、导入预览与团队日历 API。业务接口统一返回 `{data, requestId}`，错误返回 `{error, requestId}`。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /preferences:
    get:
      tags: [Preferences]
      operationId: getPreferences
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      tags: [Preferences]
      operationId: updatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Preference' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /workspaces/{workspaceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    patch:
      tags: [Workspaces]
      operationId: updateWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Workspaces]
      operationId: deleteWorkspace
      parameters:
        - name: confirm
          in: query
          required: true
          schema: { type: string }
          description: 必须与工作区名称一致。
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
...
  /workspaces/{workspaceId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/UserId'
    patch:
      tags: [Members]
      operationId: updateMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, status]
              properties:
                role: { $ref: '#/components/schemas/Role' }
                status: { $ref: '#/components/schemas/MemberStatus' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
...
  /workspaces/{workspaceId}/shifts/{shiftId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - name: shiftId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    put:
      tags: [Shifts]
      operationId: updateShift
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShiftInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /workspaces/{workspaceId}/schedules/batch:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    post:
      tags: [Schedules]
      operationId: batchSchedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ScheduleInput'
                - type: object
                  required: [userId, dates]
                  properties:
                    userId: { type: integer, format: int64 }
                    dates:
                      type: array
                      maxItems: 366
                      items: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /workspaces/{workspaceId}/audits:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Workspaces]
      operationId: listAudits
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    Error:
      description: 失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string, format: uuid }
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        requestId: { type: string, format: uuid }
...
    Preference:
      type: object
      required: [theme]
      properties:
        currentWorkspaceId: { type: [integer, 'null'], format: int64 }
        theme: { type: string, enum: [mint, sky, lilac, sakura, amber, graphite] }
    WorkspaceInput:
      type: object
      required: [name, timezone]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        timezone: { type: string, example: Asia/Shanghai }
    Role: { type: string, enum: [OWNER, ADMIN, MEMBER] }
    MemberStatus: { type: string, enum: [ACTIVE, DISABLED, REMOVED] }
    ShiftInput:
      type: object
      required: [name, code, crossDay, displayColor]
      properties:
        name: { type: string, maxLength: 80 }
        code: { type: string, maxLength: 64 }
        startTime: { type: [string, 'null'], pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        endTime: { type: [string, 'null'], pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        crossDay: { type: boolean }
        displayColor: { type: string }
        enabled: { type: boolean, default: true }
        sortOrder: { type: integer, default: 0 }
        aliases:
          type: array
          items: { type: string, maxLength: 80 }
    ScheduleSegment:
      type: object
      required: [type, crossDay]
      properties:
        type: { type: string, enum: [SHIFT, TIME_RANGE] }
        shiftId: { type: integer, format: int64 }
        startTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        endTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        crossDay: { type: boolean }
    ScheduleInput:
      type: object
      required: [status, segments]
      properties:
        status: { type: string, enum: [WORKING, REST] }
        note: { type: string, maxLength: 1000 }
        version: { type: integer, format: int64, minimum: 0 }
        segments:
          type: array
          maxItems: 8
          items: { $ref: '#/components/schemas/ScheduleSegment' }
...